Bug fix and Comment
Fixed a bug where the Disk Stats returned an error and actually returned
the word error! This caused the result to incorrectly include letters
from the error message, this has been changed to return 0's so that the
total and used size are usable in calculations.

Also added a comment about getting more info on the DNS Server for
later.